requirements.txt
numpy>=1.26,<3.0
pandas>=2.2,<3.0
scipy>=1.13,<2.0
scikit-learn>=1.5,<2.0
matplotlib>=3.9,<4.0
seaborn>=0.13,<1.0
statsmodels>=0.14,<1.0
PyYAML>=6.0,<7.0
lifetimes>=0.11,<1.0
shap>=0.47,<1.0
mlflow>=2.17,<3.0
jupyter>=1.1,<2.0
ipykernel>=6.29,<7.0
xgboost>=2.1,<3.0
lightgbm>=4.5,<5.0
catboost>=1.2,<2.0
prophet>=1.1,<2.0
pygam>=0.9,<1.0
tensorflow>=2.18,<3.0
